+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - bugfix DimensionalityEstimator dimensionality initialization
  - implement 'fixed' gaussian proces type to allow more inducing points than datapoints
  - implement `copy()` method for `Predictor` class
+ - html representation for major objects
 
 # v1.4.3